[GILJOB-86] 퀘스트 완료 api

+++ src/main/kotlin/com/yapp/giljob/domain/user/dto/response/UserInfoResponseDto.kt
@@ -2,5 +2,5 @@ package com.yapp.giljob.domain.user.dto.response
 
 class UserInfoResponseDto(
     var userId: Long,
-    var point: Int
+    var point: Long
 )
